requirements.txt
beautifulsoup4==4.5.3
aiohttp==3.3.2
path.py==11.0.1
PyYAML==3.12
Flask==1.0.2
Jinja2==2.10
lxml==3.7.3
pushbullet.py==0.11.0